LayoutTests/platform/mac/accessibility/table-scope-expected.txt

Community Courses -- Bath Autumn 1997
Course Name	Course Tutor	Summary	Code	Fee
After the Civil War	Dr. John Wroughton	 The course will examine the turbulent years in England after 1646. 6 weekly meetings starting Monday 13th October.	H27	£32	£32
An Introduction to Anglo-Saxon England	Mark Cottle	 One day course introducing the early medieval period reconstruction the Anglo-Saxons and their society. Saturday 18th October.	H28	£18	£18
The Glory that was Greece	 Birthplace of democracy, philosophy, heartland of theater, home of argument. The Romans may have done it but the Greeks did it first. Saturday day school 25th October 1997	H30	£18	£18
This tests the scope attribute on table cells will manifest as the right header elements on the cells in that row or column.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colHeaders[0].isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS colHeaders[0].isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS colHeaders[0].isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS colHeaders[1].isEqual(table.cellForColumnAndRow(1, 1)) is true
PASS rowHeaders[0].isEqual(table.cellForColumnAndRow(0, 2)) is true
PASS colHeaders[0].isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS colHeaders[1].isEqual(table.cellForColumnAndRow(3, 1)) is true
PASS rowHeaders[0].isEqual(table.cellForColumnAndRow(0, 3)) is true
PASS rowHeaders[1].isEqual(table.cellForColumnAndRow(1, 3)) is true
PASS rowHeaders[0].isEqual(table.cellForColumnAndRow(0, 4)) is true
PASS rowHeaders[1].isEqual(table.cellForColumnAndRow(1, 3)) is true
PASS colHeaders[0].isEqual(table.cellForColumnAndRow(0, 0)) is true
PASS colHeaders[1].isEqual(table.cellForColumnAndRow(4, 1)) is true
PASS successfullyParsed is true

TEST COMPLETE
